Handling non numerical data, training and evaluate
We create some costum transformers in order to handle the no numerical
data, created pipelines in order tu use this transformers, as well as
OneHotEncoding, finally we join all the data using a ColumnTransformer
with or pipelines and costum transformers.

Once the preprocess of the data were done, we trained and evaluated some
models (using cross validation)

Next we want to Fine tune our better model and use it to do predictions
using test dataset.